msm: camera: icp: Change to dump ICP pagefault data properly
ICP context bank is shared between multiple ICPs based on IPE and
OFE connected to them. In case of pagefault an optimization is
performed to avoid querying the same information on both the nodes.
However, at first query is done to get the corresponding node that
have caused the fault, due to logical error, the same information
was not getting populated correctly.
This was because node contains the context on which the fault has
occurred and the pagefault handler should be iterating on all
context to gather right information rather than only on the zeroth
context.
This change makes sure that we iterate over all context to find the
right node on which the pid is matching and then dumps the request
configuration which caused the pagefault.

CRs-Fixed: 4115493
Change-Id: Ibecc1aee3bf7b76fd96b4993049fae5b01e69fc0
Signed-off-by: Abhilash Kumar <quic_krabhi@quicinc.com>