qcacld-3.0: Fix SAP+SAP failing to start on the same channel
In non-DBS HW, currently only STA+SAP is considered for SCC, while
SAP+SAP SCC will be mistakenly treated as MCC, resulting in SAP+SAP
failing to start on the same channel.

When determining whether two vdev are on the same channel SCC, the
case of SAP+SAP/GO should also be taken into account to fix this
issue.

Change-Id: I96d54831d271704c8060812afd4da74aab3fce4e
CRs-Fixed: 4084685