dsp-kernel: Add check for map attr type to avoid null ptr dereference
When a map is created with the FASTRPC_ATTR_NOMAP attribute
but without the FASTRPC_MAP_FD_NOMAP flag, map smmucb is NULL.
When freeing this map, only the map flags field is checked.
This causes a NULL pointer dereference when accessing
smmucb map_mutex. Add a check for map attr.

Change-Id: I4f08be23673f095552782e07b06de05634cac065
Signed-off-by: Om Deore <quic_odeore@quicinc.com>