disp: msm: dsi: fix FPS switch command transfer logic
Currently, the FPS switch command is sent during prepare_commit, at
which point the panel does not yet have information about the new mode.
As a result, the command is fetched based on the previous FPS setting.
To address this, move the FPS switch command to pre_kickoff, where the
panel mode has already been updated to the new one.

Change-Id: I6b7594d5a7ab8bf41041e308a3652d51436e4c1d
Signed-off-by: Srihitha Tangudu <quic_tangudu@quicinc.com>