qcacld-3.0: Preserve WMM parameters during link switch
WMM parameters are cleared for every disconnect including the
link switch disconnect case. This causes the traffic sent in
the window of teardown of the partner link and the activation
of the standby link to be transmitted on the default BK tid,
instead of the configured AC,leading to reduced QoS performance.

To fix this, don't clear of the WMM params of the adapter for
link switch disconnects.

Change-Id: Ic140c53cf596930d438aaa86aaf7de9301ff56d4
CRs-Fixed: 4116023