qcacmn: Increase timer threshold for RX exception ring interrupt
Currently, timer threshold for RX exception ring interrupt is configured
as 8 usecs. If the NAPI processing get delayed for some reason, then WLAN
hardware will keep firing interrupts every 8usecs and in the worst-case
scenario, this would result in an interrupt storm impacting other use cases
in the system.

Increase the timer threshold to 512 usecs to address this problem.

Change-Id: I7343232da1628f0f7a26dff5c4f1304e5f103804
CRs-Fixed: 4167656