kgsl: Fix null pointer dereference during secure memory alloc
Currently, when we attempt to allocate shared memory we track the
underlying pages in the memdesc with a scatter gather table. Due
to this we do not assign a pointer to the memdesc->pages.

Under memory pressure, allocation of the sg or sgt may fail which
is detected properly. However, our clean up code drops the page
references based on memdesc->pages. Since secure memory does not
track via memdesc->pages we get a NPD. We'll quickly assign pages
before calling _kgsl_free_pages, and modify _kgsl_free_pages to
clean up the references/count after.

Change-Id: I763634bf00ebef7a57a8d56c21ee8d45cf2b39da
Signed-off-by: Lynus Vaz <quic_lvaz@quicinc.com>
Signed-off-by: Pankaj Gupta <gpankaj@qti.qualcomm.com>